// src/types/invoices/model/invoice.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zero consult invoicing - OpenAPI 3.2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InvoiceStatus } from './invoice-status';

export interface Invoice {
    'customerId': string;
    'expireDate': string;
    'id'?: string;
    'invoiceDate': string;
    'invoiceFile'?: string;
    'invoiceNumber'?: number;
    'invoiceUntil': string;
    'notes': string;
    'status': InvoiceStatus;
    'totalWithoutVat': number;
    'vatPercent': number;
}